+++ setup.cfg
@@ -3,6 +3,7 @@
 # for now, just check the main code
 filename =
     clifford/__init__.py
+    clifford/_*.py
 
 ignore =
     # flake8 defaults
